+++ lib/proof/helper.js
@@ -55,7 +55,6 @@ ProofHelperKlass.prototype = {
     let accountProofInstance = new AccountProof(stateRoot, db);
     let accountProof = await accountProofInstance.perform(contractAddress);
     let accountValue = accountProof.toHash().data.value;
-    console.log("account proof storage root", accountValue);
     let decodedValue = ethUtils.rlp.decode('0x' + accountValue);
     return '0x' + decodedValue[2].toString('hex');
   }
