+++ test/integration/services/balance_branded_token.js
@@ -6,25 +6,26 @@ const chai = require('chai')
 ;
 
 // Load cache service
-const rootPrefix = "../.."
+const rootPrefix = "../../.."
   , openstPlatform = require(rootPrefix + '/index')
   , platformServices = openstPlatform.services.balance
   , web3ProviderFactory = require(rootPrefix + '/lib/web3/providers/factory')
   , brandedTokenConfigPath = os.homedir() + "/openst-setup/branded_tokens.json"
   , brandedTokenConfig = require(brandedTokenConfigPath)
 ;
-
+var brandedTokenDetails = brandedTokenConfig[Object.keys(brandedTokenConfig)[0]]
 var testValidData = {
-  address: process.env.OST_UTILITY_CHAIN_OWNER_ADDR
+  address: brandedTokenDetails['Reserve'],
+  erc20_address: brandedTokenDetails['ERC20']
 };
 
-describe('services/balance/simple_token_prime', function () {
+describe('services/balance/branded_token', function () {
 
   it('should return promise', async function () {
     var dupData = JSON.parse(JSON.stringify(testValidData));
 
-    var simpleTokenPrimeObj = new platformServices.simpleTokenPrime(dupData)
-      , response = simpleTokenPrimeObj.perform()
+    var brandedTokenObj = new platformServices.brandedToken(dupData)
+      , response = brandedTokenObj.perform()
     ;
     assert.typeOf(response, 'Promise');
   });
@@ -32,8 +33,8 @@ describe('services/balance/simple_token_prime', function () {
   it('should fail when params is undefined', async function () {
     var dupData = undefined;
 
-    var simpleTokenPrimeObj = new platformServices.simpleTokenPrime(dupData)
-      , response = await simpleTokenPrimeObj.perform()
+    var brandedTokenObj = new platformServices.brandedToken(dupData)
+      , response = await brandedTokenObj.perform()
     ;
     assert.equal(response.isSuccess(), false);
   });
@@ -41,8 +42,8 @@ describe('services/balance/simple_token_prime', function () {
   it('should fail when params is a string', async function () {
     var dupData = 'abc';
 
-    var simpleTokenPrimeObj = new platformServices.simpleTokenPrime(dupData)
-      , response = await simpleTokenPrimeObj.perform()
+    var brandedTokenObj = new platformServices.brandedToken(dupData)
+      , response = await brandedTokenObj.perform()
     ;
     assert.equal(response.isSuccess(), false);
   });
@@ -50,8 +51,8 @@ describe('services/balance/simple_token_prime', function () {
   it('should fail when params is empty object', async function () {
     var dupData = {};
 
-    var simpleTokenPrimeObj = new platformServices.simpleTokenPrime(dupData)
-      , response = await simpleTokenPrimeObj.perform()
+    var brandedTokenObj = new platformServices.brandedToken(dupData)
+      , response = await brandedTokenObj.perform()
     ;
     assert.equal(response.isSuccess(), false);
   });
@@ -59,8 +60,8 @@ describe('services/balance/simple_token_prime', function () {
   it('should fail when params is empty array', async function () {
     var dupData = [];
 
-    var simpleTokenPrimeObj = new platformServices.simpleTokenPrime(dupData)
-      , response = await simpleTokenPrimeObj.perform()
+    var brandedTokenObj = new platformServices.brandedToken(dupData)
+      , response = await brandedTokenObj.perform()
     ;
     assert.equal(response.isSuccess(), false);
   });
@@ -70,8 +71,8 @@ describe('services/balance/simple_token_prime', function () {
     var dupData = JSON.parse(JSON.stringify(testValidData));
     dupData.address = '';
 
-    var simpleTokenPrimeObj = new platformServices.simpleTokenPrime(dupData)
-      , response = await simpleTokenPrimeObj.perform()
+    var brandedTokenObj = new platformServices.brandedToken(dupData)
+      , response = await brandedTokenObj.perform()
     ;
     assert.equal(response.isSuccess(), false);
   });
@@ -80,8 +81,29 @@ describe('services/balance/simple_token_prime', function () {
     var dupData = JSON.parse(JSON.stringify(testValidData));
     dupData.address = '0xh32323';
 
-    var simpleTokenPrimeObj = new platformServices.simpleTokenPrime(dupData)
-      , response = await simpleTokenPrimeObj.perform()
+    var brandedTokenObj = new platformServices.brandedToken(dupData)
+      , response = await brandedTokenObj.perform()
+    ;
+    assert.equal(response.isSuccess(), false);
+  });
+
+  // ERC20 Variations
+  it('should fail when erc20 address is blank', async function () {
+    var dupData = JSON.parse(JSON.stringify(testValidData));
+    dupData.erc20_address = '';
+
+    var brandedTokenObj = new platformServices.brandedToken(dupData)
+      , response = await brandedTokenObj.perform()
+    ;
+    assert.equal(response.isSuccess(), false);
+  });
+
+  it('should fail when erc20 address is invalid', async function () {
+    var dupData = JSON.parse(JSON.stringify(testValidData));
+    dupData.erc20_address = '0x2323';
+
+    var brandedTokenObj = new platformServices.brandedToken(dupData)
+      , response = await brandedTokenObj.perform()
     ;
     assert.equal(response.isSuccess(), false);
   });
@@ -91,8 +113,8 @@ describe('services/balance/simple_token_prime', function () {
   it('should pass when everything is valid', async function () {
     var dupData = JSON.parse(JSON.stringify(testValidData));
 
-    var simpleTokenPrimeObj = new platformServices.simpleTokenPrime(dupData)
-      , response = await simpleTokenPrimeObj.perform()
+    var brandedTokenObj = new platformServices.brandedToken(dupData)
+      , response = await brandedTokenObj.perform()
     ;
     assert.equal(response.isSuccess(), true);
     assert.isNotNull(response.data.balance);
