+++ test/integration/services/balance_simple_token.js
@@ -6,26 +6,25 @@ const chai = require('chai')
 ;
 
 // Load cache service
-const rootPrefix = "../.."
+const rootPrefix = "../../.."
   , openstPlatform = require(rootPrefix + '/index')
   , platformServices = openstPlatform.services.balance
   , web3ProviderFactory = require(rootPrefix + '/lib/web3/providers/factory')
   , brandedTokenConfigPath = os.homedir() + "/openst-setup/branded_tokens.json"
   , brandedTokenConfig = require(brandedTokenConfigPath)
 ;
-var brandedTokenDetails = brandedTokenConfig[Object.keys(brandedTokenConfig)[0]]
+
 var testValidData = {
-  address: brandedTokenDetails['Reserve'],
-  erc20_address: brandedTokenDetails['ERC20']
+  address: process.env.OST_UTILITY_CHAIN_OWNER_ADDR
 };
 
-describe('services/balance/branded_token', function () {
+describe('services/balance/simple_token', function () {
 
   it('should return promise', async function () {
     var dupData = JSON.parse(JSON.stringify(testValidData));
 
-    var brandedTokenObj = new platformServices.brandedToken(dupData)
-      , response = brandedTokenObj.perform()
+    var simpleTokenObj = new platformServices.simpleToken(dupData)
+      , response = simpleTokenObj.perform()
     ;
     assert.typeOf(response, 'Promise');
   });
@@ -33,8 +32,8 @@ describe('services/balance/branded_token', function () {
   it('should fail when params is undefined', async function () {
     var dupData = undefined;
 
-    var brandedTokenObj = new platformServices.brandedToken(dupData)
-      , response = await brandedTokenObj.perform()
+    var simpleTokenObj = new platformServices.simpleToken(dupData)
+      , response = await simpleTokenObj.perform()
     ;
     assert.equal(response.isSuccess(), false);
   });
@@ -42,8 +41,8 @@ describe('services/balance/branded_token', function () {
   it('should fail when params is a string', async function () {
     var dupData = 'abc';
 
-    var brandedTokenObj = new platformServices.brandedToken(dupData)
-      , response = await brandedTokenObj.perform()
+    var simpleTokenObj = new platformServices.simpleToken(dupData)
+      , response = await simpleTokenObj.perform()
     ;
     assert.equal(response.isSuccess(), false);
   });
@@ -51,8 +50,8 @@ describe('services/balance/branded_token', function () {
   it('should fail when params is empty object', async function () {
     var dupData = {};
 
-    var brandedTokenObj = new platformServices.brandedToken(dupData)
-      , response = await brandedTokenObj.perform()
+    var simpleTokenObj = new platformServices.simpleToken(dupData)
+      , response = await simpleTokenObj.perform()
     ;
     assert.equal(response.isSuccess(), false);
   });
@@ -60,8 +59,8 @@ describe('services/balance/branded_token', function () {
   it('should fail when params is empty array', async function () {
     var dupData = [];
 
-    var brandedTokenObj = new platformServices.brandedToken(dupData)
-      , response = await brandedTokenObj.perform()
+    var simpleTokenObj = new platformServices.simpleToken(dupData)
+      , response = await simpleTokenObj.perform()
     ;
     assert.equal(response.isSuccess(), false);
   });
@@ -71,8 +70,8 @@ describe('services/balance/branded_token', function () {
     var dupData = JSON.parse(JSON.stringify(testValidData));
     dupData.address = '';
 
-    var brandedTokenObj = new platformServices.brandedToken(dupData)
-      , response = await brandedTokenObj.perform()
+    var simpleTokenObj = new platformServices.simpleToken(dupData)
+      , response = await simpleTokenObj.perform()
     ;
     assert.equal(response.isSuccess(), false);
   });
@@ -81,29 +80,8 @@ describe('services/balance/branded_token', function () {
     var dupData = JSON.parse(JSON.stringify(testValidData));
     dupData.address = '0xh32323';
 
-    var brandedTokenObj = new platformServices.brandedToken(dupData)
-      , response = await brandedTokenObj.perform()
-    ;
-    assert.equal(response.isSuccess(), false);
-  });
-
-  // ERC20 Variations
-  it('should fail when erc20 address is blank', async function () {
-    var dupData = JSON.parse(JSON.stringify(testValidData));
-    dupData.erc20_address = '';
-
-    var brandedTokenObj = new platformServices.brandedToken(dupData)
-      , response = await brandedTokenObj.perform()
-    ;
-    assert.equal(response.isSuccess(), false);
-  });
-
-  it('should fail when erc20 address is invalid', async function () {
-    var dupData = JSON.parse(JSON.stringify(testValidData));
-    dupData.erc20_address = '0x2323';
-
-    var brandedTokenObj = new platformServices.brandedToken(dupData)
-      , response = await brandedTokenObj.perform()
+    var simpleTokenObj = new platformServices.simpleToken(dupData)
+      , response = await simpleTokenObj.perform()
     ;
     assert.equal(response.isSuccess(), false);
   });
@@ -113,8 +91,8 @@ describe('services/balance/branded_token', function () {
   it('should pass when everything is valid', async function () {
     var dupData = JSON.parse(JSON.stringify(testValidData));
 
-    var brandedTokenObj = new platformServices.brandedToken(dupData)
-      , response = await brandedTokenObj.perform()
+    var simpleTokenObj = new platformServices.simpleToken(dupData)
+      , response = await simpleTokenObj.perform()
     ;
     assert.equal(response.isSuccess(), true);
     assert.isNotNull(response.data.balance);
