+++ test/integration/services/balance_simple_token_prime.js
@@ -6,7 +6,7 @@ const chai = require('chai')
 ;
 
 // Load cache service
-const rootPrefix = "../.."
+const rootPrefix = "../../.."
   , openstPlatform = require(rootPrefix + '/index')
   , platformServices = openstPlatform.services.balance
   , web3ProviderFactory = require(rootPrefix + '/lib/web3/providers/factory')
@@ -18,13 +18,13 @@ var testValidData = {
   address: process.env.OST_UTILITY_CHAIN_OWNER_ADDR
 };
 
-describe('services/balance/simple_token', function () {
+describe('services/balance/simple_token_prime', function () {
 
   it('should return promise', async function () {
     var dupData = JSON.parse(JSON.stringify(testValidData));
 
-    var simpleTokenObj = new platformServices.simpleToken(dupData)
-      , response = simpleTokenObj.perform()
+    var simpleTokenPrimeObj = new platformServices.simpleTokenPrime(dupData)
+      , response = simpleTokenPrimeObj.perform()
     ;
     assert.typeOf(response, 'Promise');
   });
@@ -32,8 +32,8 @@ describe('services/balance/simple_token', function () {
   it('should fail when params is undefined', async function () {
     var dupData = undefined;
 
-    var simpleTokenObj = new platformServices.simpleToken(dupData)
-      , response = await simpleTokenObj.perform()
+    var simpleTokenPrimeObj = new platformServices.simpleTokenPrime(dupData)
+      , response = await simpleTokenPrimeObj.perform()
     ;
     assert.equal(response.isSuccess(), false);
   });
@@ -41,8 +41,8 @@ describe('services/balance/simple_token', function () {
   it('should fail when params is a string', async function () {
     var dupData = 'abc';
 
-    var simpleTokenObj = new platformServices.simpleToken(dupData)
-      , response = await simpleTokenObj.perform()
+    var simpleTokenPrimeObj = new platformServices.simpleTokenPrime(dupData)
+      , response = await simpleTokenPrimeObj.perform()
     ;
     assert.equal(response.isSuccess(), false);
   });
@@ -50,8 +50,8 @@ describe('services/balance/simple_token', function () {
   it('should fail when params is empty object', async function () {
     var dupData = {};
 
-    var simpleTokenObj = new platformServices.simpleToken(dupData)
-      , response = await simpleTokenObj.perform()
+    var simpleTokenPrimeObj = new platformServices.simpleTokenPrime(dupData)
+      , response = await simpleTokenPrimeObj.perform()
     ;
     assert.equal(response.isSuccess(), false);
   });
@@ -59,8 +59,8 @@ describe('services/balance/simple_token', function () {
   it('should fail when params is empty array', async function () {
     var dupData = [];
 
-    var simpleTokenObj = new platformServices.simpleToken(dupData)
-      , response = await simpleTokenObj.perform()
+    var simpleTokenPrimeObj = new platformServices.simpleTokenPrime(dupData)
+      , response = await simpleTokenPrimeObj.perform()
     ;
     assert.equal(response.isSuccess(), false);
   });
@@ -70,8 +70,8 @@ describe('services/balance/simple_token', function () {
     var dupData = JSON.parse(JSON.stringify(testValidData));
     dupData.address = '';
 
-    var simpleTokenObj = new platformServices.simpleToken(dupData)
-      , response = await simpleTokenObj.perform()
+    var simpleTokenPrimeObj = new platformServices.simpleTokenPrime(dupData)
+      , response = await simpleTokenPrimeObj.perform()
     ;
     assert.equal(response.isSuccess(), false);
   });
@@ -80,8 +80,8 @@ describe('services/balance/simple_token', function () {
     var dupData = JSON.parse(JSON.stringify(testValidData));
     dupData.address = '0xh32323';
 
-    var simpleTokenObj = new platformServices.simpleToken(dupData)
-      , response = await simpleTokenObj.perform()
+    var simpleTokenPrimeObj = new platformServices.simpleTokenPrime(dupData)
+      , response = await simpleTokenPrimeObj.perform()
     ;
     assert.equal(response.isSuccess(), false);
   });
@@ -91,8 +91,8 @@ describe('services/balance/simple_token', function () {
   it('should pass when everything is valid', async function () {
     var dupData = JSON.parse(JSON.stringify(testValidData));
 
-    var simpleTokenObj = new platformServices.simpleToken(dupData)
-      , response = await simpleTokenObj.perform()
+    var simpleTokenPrimeObj = new platformServices.simpleTokenPrime(dupData)
+      , response = await simpleTokenPrimeObj.perform()
     ;
     assert.equal(response.isSuccess(), true);
     assert.isNotNull(response.data.balance);
