+++ test/integration/services/transaction_transfer_branded_token.js
@@ -6,7 +6,7 @@ const chai = require('chai')
 ;
 
 // Load cache service
-const rootPrefix = "../.."
+const rootPrefix = "../../.."
   , openstPlatform = require(rootPrefix + '/index')
   , platformServices = openstPlatform.services.transaction.transfer
   , web3ProviderFactory = require(rootPrefix + '/lib/web3/providers/factory')
@@ -14,35 +14,69 @@ const rootPrefix = "../.."
   , brandedTokenConfig = require(brandedTokenConfigPath)
 ;
 
+var brandedTokenDetails = brandedTokenConfig[Object.keys(brandedTokenConfig)[0]]
+
 var testValidData = {
-  sender_address: process.env.OST_FOUNDATION_ADDR,
-  sender_passphrase: process.env.OST_FOUNDATION_PASSPHRASE,
-  sender_name: 'foundation',
-  recipient_address: process.env.OST_UTILITY_CHAIN_OWNER_ADDR,
-  recipient_name: 'utilityChainOwner',
-  amount_in_wei: 2000,
+  erc20_address: brandedTokenDetails['ERC20'],
+  sender_address: process.env.OST_UTILITY_CHAIN_OWNER_ADDR,
+  sender_passphrase: process.env.OST_UTILITY_CHAIN_OWNER_PASSPHRASE,
+  sender_name: 'utilityChainOwner',
+  recipient_address: process.env.OST_STAKER_ADDR,
+  recipient_name: 'staker',
+  amount_in_wei: 2,
   options: {
     returnType: 'txHash',
-    tag: 'GasRefill'
+    tag: 'ILoveOST'
   }
 };
 
-describe('services/transaction/transfer/eth', function () {
+describe('services/transaction/transfer/branded_token', function () {
 
   it('should return promise', async function () {
     var dupData = JSON.parse(JSON.stringify(testValidData));
 
-    var ethObj = new platformServices.eth(dupData)
-      , response = ethObj.perform()
+    var brandedTokenObj = new platformServices.brandedToken(dupData)
+      , response = brandedTokenObj.perform()
     ;
     assert.typeOf(response, 'Promise');
   });
 
+  // Prefil BT to sender
+
+  it('should pass prefil BT to sender promise', async function () {
+    // refill gas
+    var serviceObj = new openstPlatform.services.transaction.transfer.simpleTokenPrime(
+      {
+        sender_name: 'utilityChainOwner', recipient_address: brandedTokenDetails['Reserve'],
+        amount_in_wei: 1000000000000000000, options: {returnType: 'txReceipt', tag: 'GasRefill'}
+      });
+    await serviceObj.perform();
+
+    var dupData = JSON.parse(JSON.stringify(testValidData));
+    dupData.options.returnType = 'txReceipt';
+    dupData.sender_name = '';
+    dupData.sender_address = brandedTokenDetails['Reserve'];
+    dupData.sender_passphrase = brandedTokenDetails['ReservePassphrase'];
+    dupData.recipient_name = 'utilityChainOwner';
+    dupData.amount_in_wei = 100;
+
+
+    var brandedTokenObj = new platformServices.brandedToken(dupData)
+      , response = await brandedTokenObj.perform()
+    ;
+    assert.equal(response.isSuccess(), true);
+    assert.isNotNull(response.data.transaction_uuid);
+    assert.isNotNull(response.data.transaction_hash);
+    assert.isNumber(response.data.transaction_receipt.blockNumber);
+  });
+
+  // Params validations
+
   it('should fail when params is undefined', async function () {
     var dupData = undefined;
 
-    var ethObj = new platformServices.eth(dupData)
-      , response = await ethObj.perform()
+    var brandedTokenObj = new platformServices.brandedToken(dupData)
+      , response = await brandedTokenObj.perform()
     ;
     assert.equal(response.isSuccess(), false);
   });
@@ -50,8 +84,8 @@ describe('services/transaction/transfer/eth', function () {
   it('should fail when params is a string', async function () {
     var dupData = 'abc';
 
-    var ethObj = new platformServices.eth(dupData)
-      , response = await ethObj.perform()
+    var brandedTokenObj = new platformServices.brandedToken(dupData)
+      , response = await brandedTokenObj.perform()
     ;
     assert.equal(response.isSuccess(), false);
   });
@@ -59,8 +93,8 @@ describe('services/transaction/transfer/eth', function () {
   it('should fail when params is empty object', async function () {
     var dupData = {};
 
-    var ethObj = new platformServices.eth(dupData)
-      , response = await ethObj.perform()
+    var brandedTokenObj = new platformServices.brandedToken(dupData)
+      , response = await brandedTokenObj.perform()
     ;
     assert.equal(response.isSuccess(), false);
   });
@@ -68,8 +102,8 @@ describe('services/transaction/transfer/eth', function () {
   it('should fail when params is empty array', async function () {
     var dupData = [];
 
-    var ethObj = new platformServices.eth(dupData)
-      , response = await ethObj.perform()
+    var brandedTokenObj = new platformServices.brandedToken(dupData)
+      , response = await brandedTokenObj.perform()
     ;
     assert.equal(response.isSuccess(), false);
   });
@@ -78,8 +112,8 @@ describe('services/transaction/transfer/eth', function () {
     var dupData = JSON.parse(JSON.stringify(testValidData));
     dupData.options.tag = 'a@b';
 
-    var ethObj = new platformServices.eth(dupData)
-      , response = await ethObj.perform()
+    var brandedTokenObj = new platformServices.brandedToken(dupData)
+      , response = await brandedTokenObj.perform()
     ;
     assert.equal(response.isSuccess(), false);
   });
@@ -90,8 +124,8 @@ describe('services/transaction/transfer/eth', function () {
     var dupData = JSON.parse(JSON.stringify(testValidData));
     dupData.sender_name = 'Google';
 
-    var ethObj = new platformServices.eth(dupData)
-      , response = await ethObj.perform()
+    var brandedTokenObj = new platformServices.brandedToken(dupData)
+      , response = await brandedTokenObj.perform()
     ;
     assert.equal(response.isSuccess(), false);
   });
@@ -101,8 +135,8 @@ describe('services/transaction/transfer/eth', function () {
     dupData.sender_name = ''; // has higher priority
     dupData.sender_address = 'abc';
 
-    var ethObj = new platformServices.eth(dupData)
-      , response = await ethObj.perform()
+    var brandedTokenObj = new platformServices.brandedToken(dupData)
+      , response = await brandedTokenObj.perform()
     ;
     assert.equal(response.isSuccess(), false);
   });
@@ -110,10 +144,10 @@ describe('services/transaction/transfer/eth', function () {
   it('should fail when sender address is valid, but has no balance', async function () {
     var dupData = JSON.parse(JSON.stringify(testValidData));
     dupData.sender_name = ''; // has higher priority
-    dupData.sender_address = '0xb4d7bedf714e6c7cd1a641f705870fa19144a021';
+    dupData.sender_address = '0xb4d7bedf714e6c7cd1a641f705870fa19144a061';
 
-    var ethObj = new platformServices.eth(dupData)
-      , response = await ethObj.perform()
+    var brandedTokenObj = new platformServices.brandedToken(dupData)
+      , response = await brandedTokenObj.perform()
     ;
     assert.equal(response.isSuccess(), false);
   });
@@ -123,8 +157,8 @@ describe('services/transaction/transfer/eth', function () {
     dupData.sender_name = ''; // has higher priority
     dupData.sender_passphrase = 'abc';
 
-    var ethObj = new platformServices.eth(dupData)
-      , response = await ethObj.perform()
+    var brandedTokenObj = new platformServices.brandedToken(dupData)
+      , response = await brandedTokenObj.perform()
     ;
     assert.equal(response.isSuccess(), false);
   });
@@ -136,8 +170,8 @@ describe('services/transaction/transfer/eth', function () {
     dupData.recipient_name = ''; // has higher priority
     dupData.recipient_address = 'abc';
 
-    var ethObj = new platformServices.eth(dupData)
-      , response = await ethObj.perform()
+    var brandedTokenObj = new platformServices.brandedToken(dupData)
+      , response = await brandedTokenObj.perform()
     ;
     assert.equal(response.isSuccess(), false);
   });
@@ -146,8 +180,8 @@ describe('services/transaction/transfer/eth', function () {
     var dupData = JSON.parse(JSON.stringify(testValidData));
     dupData.recipient_name = 'abc';
 
-    var ethObj = new platformServices.eth(dupData)
-      , response = await ethObj.perform()
+    var brandedTokenObj = new platformServices.brandedToken(dupData)
+      , response = await brandedTokenObj.perform()
     ;
     assert.equal(response.isSuccess(), false);
   });
@@ -158,8 +192,8 @@ describe('services/transaction/transfer/eth', function () {
     var dupData = JSON.parse(JSON.stringify(testValidData));
     dupData.amount_in_wei = undefined;
 
-    var ethObj = new platformServices.eth(dupData)
-      , response = await ethObj.perform()
+    var brandedTokenObj = new platformServices.brandedToken(dupData)
+      , response = await brandedTokenObj.perform()
     ;
     assert.equal(response.isSuccess(), false);
   });
@@ -168,8 +202,8 @@ describe('services/transaction/transfer/eth', function () {
     var dupData = JSON.parse(JSON.stringify(testValidData));
     dupData.amount_in_wei = 'abc';
 
-    var ethObj = new platformServices.eth(dupData)
-      , response = await ethObj.perform()
+    var brandedTokenObj = new platformServices.brandedToken(dupData)
+      , response = await brandedTokenObj.perform()
     ;
     assert.equal(response.isSuccess(), false);
   });
@@ -178,8 +212,8 @@ describe('services/transaction/transfer/eth', function () {
     var dupData = JSON.parse(JSON.stringify(testValidData));
     dupData.amount_in_wei = 100.2;
 
-    var ethObj = new platformServices.eth(dupData)
-      , response = await ethObj.perform()
+    var brandedTokenObj = new platformServices.brandedToken(dupData)
+      , response = await brandedTokenObj.perform()
     ;
     assert.equal(response.isSuccess(), false);
   });
@@ -188,8 +222,8 @@ describe('services/transaction/transfer/eth', function () {
     var dupData = JSON.parse(JSON.stringify(testValidData));
     dupData.amount_in_wei = 0.2;
 
-    var ethObj = new platformServices.eth(dupData)
-      , response = await ethObj.perform()
+    var brandedTokenObj = new platformServices.brandedToken(dupData)
+      , response = await brandedTokenObj.perform()
     ;
     assert.equal(response.isSuccess(), false);
   });
@@ -198,8 +232,8 @@ describe('services/transaction/transfer/eth', function () {
     var dupData = JSON.parse(JSON.stringify(testValidData));
     dupData.amount_in_wei = 0;
 
-    var ethObj = new platformServices.eth(dupData)
-      , response = await ethObj.perform()
+    var brandedTokenObj = new platformServices.brandedToken(dupData)
+      , response = await brandedTokenObj.perform()
     ;
     assert.equal(response.isSuccess(), false);
   });
@@ -208,8 +242,8 @@ describe('services/transaction/transfer/eth', function () {
     var dupData = JSON.parse(JSON.stringify(testValidData));
     dupData.amount_in_wei = -100;
 
-    var ethObj = new platformServices.eth(dupData)
-      , response = await ethObj.perform()
+    var brandedTokenObj = new platformServices.brandedToken(dupData)
+      , response = await brandedTokenObj.perform()
     ;
     assert.equal(response.isSuccess(), false);
   });
@@ -219,8 +253,8 @@ describe('services/transaction/transfer/eth', function () {
   it('should pass when everything is valid', async function () {
     var dupData = JSON.parse(JSON.stringify(testValidData));
 
-    var ethObj = new platformServices.eth(dupData)
-      , response = await ethObj.perform()
+    var brandedTokenObj = new platformServices.brandedToken(dupData)
+      , response = await brandedTokenObj.perform()
     ;
     assert.equal(response.isSuccess(), true);
     assert.isNotNull(response.data.transaction_uuid);
@@ -231,8 +265,8 @@ describe('services/transaction/transfer/eth', function () {
     var dupData = JSON.parse(JSON.stringify(testValidData));
     dupData.options.returnType = 'myReturnType';
 
-    var ethObj = new platformServices.eth(dupData)
-      , response = await ethObj.perform()
+    var brandedTokenObj = new platformServices.brandedToken(dupData)
+      , response = await brandedTokenObj.perform()
     ;
 
     assert.equal(response.isSuccess(), true);
@@ -245,8 +279,8 @@ describe('services/transaction/transfer/eth', function () {
     var dupData = JSON.parse(JSON.stringify(testValidData));
     dupData.options.returnType = 'uuid';
 
-    var ethObj = new platformServices.eth(dupData)
-      , response = await ethObj.perform()
+    var brandedTokenObj = new platformServices.brandedToken(dupData)
+      , response = await brandedTokenObj.perform()
     ;
 
     assert.equal(response.isSuccess(), true);
@@ -259,8 +293,8 @@ describe('services/transaction/transfer/eth', function () {
     var dupData = JSON.parse(JSON.stringify(testValidData));
     dupData.options.returnType = 'txReceipt';
 
-    var ethObj = new platformServices.eth(dupData)
-      , response = await ethObj.perform()
+    var brandedTokenObj = new platformServices.brandedToken(dupData)
+      , response = await brandedTokenObj.perform()
     ;
 
     assert.equal(response.isSuccess(), true);
@@ -268,5 +302,8 @@ describe('services/transaction/transfer/eth', function () {
     assert.isNotNull(response.data.transaction_hash);
     assert.isNumber(response.data.transaction_receipt.blockNumber);
   });
+
+  //TODO: check balance in cache
+
 
 });
