+++ test/integration/services/transaction_transfer_simple_token_prime.js
@@ -6,7 +6,7 @@ const chai = require('chai')
 ;
 
 // Load cache service
-const rootPrefix = "../.."
+const rootPrefix = "../../.."
   , openstPlatform = require(rootPrefix + '/index')
   , platformServices = openstPlatform.services.transaction.transfer
   , web3ProviderFactory = require(rootPrefix + '/lib/web3/providers/factory')
@@ -15,25 +15,25 @@ const rootPrefix = "../.."
 ;
 
 var testValidData = {
-  sender_address: process.env.OST_FOUNDATION_ADDR,
-  sender_passphrase: process.env.OST_FOUNDATION_PASSPHRASE,
-  sender_name: 'foundation',
-  recipient_address: process.env.OST_UTILITY_CHAIN_OWNER_ADDR,
-  recipient_name: 'utilityChainOwner',
+  sender_address: process.env.OST_UTILITY_CHAIN_OWNER_ADDR,
+  sender_passphrase: process.env.OST_UTILITY_CHAIN_OWNER_PASSPHRASE,
+  sender_name: 'utilityChainOwner',
+  recipient_address: process.env.OST_STAKER_ADDR,
+  recipient_name: 'staker',
   amount_in_wei: 20,
   options: {
     returnType: 'txHash',
-    tag: 'Grant'
+    tag: 'GasRefill'
   }
 };
 
-describe('services/transaction/transfer/simple_token', function () {
+describe('services/transaction/transfer/simple_token_prime', function () {
 
   it('should return promise', async function () {
     var dupData = JSON.parse(JSON.stringify(testValidData));
 
-    var simpleTokenObj = new platformServices.simpleToken(dupData)
-      , response = simpleTokenObj.perform()
+    var simpleTokenPrimeObj = new platformServices.simpleTokenPrime(dupData)
+      , response = simpleTokenPrimeObj.perform()
     ;
     assert.typeOf(response, 'Promise');
   });
@@ -41,8 +41,8 @@ describe('services/transaction/transfer/simple_token', function () {
   it('should fail when params is undefined', async function () {
     var dupData = undefined;
 
-    var simpleTokenObj = new platformServices.simpleToken(dupData)
-      , response = await simpleTokenObj.perform()
+    var simpleTokenPrimeObj = new platformServices.simpleTokenPrime(dupData)
+      , response = await simpleTokenPrimeObj.perform()
     ;
     assert.equal(response.isSuccess(), false);
   });
@@ -50,8 +50,8 @@ describe('services/transaction/transfer/simple_token', function () {
   it('should fail when params is a string', async function () {
     var dupData = 'abc';
 
-    var simpleTokenObj = new platformServices.simpleToken(dupData)
-      , response = await simpleTokenObj.perform()
+    var simpleTokenPrimeObj = new platformServices.simpleTokenPrime(dupData)
+      , response = await simpleTokenPrimeObj.perform()
     ;
     assert.equal(response.isSuccess(), false);
   });
@@ -59,8 +59,8 @@ describe('services/transaction/transfer/simple_token', function () {
   it('should fail when params is empty object', async function () {
     var dupData = {};
 
-    var simpleTokenObj = new platformServices.simpleToken(dupData)
-      , response = await simpleTokenObj.perform()
+    var simpleTokenPrimeObj = new platformServices.simpleTokenPrime(dupData)
+      , response = await simpleTokenPrimeObj.perform()
     ;
     assert.equal(response.isSuccess(), false);
   });
@@ -68,8 +68,8 @@ describe('services/transaction/transfer/simple_token', function () {
   it('should fail when params is empty array', async function () {
     var dupData = [];
 
-    var simpleTokenObj = new platformServices.simpleToken(dupData)
-      , response = await simpleTokenObj.perform()
+    var simpleTokenPrimeObj = new platformServices.simpleTokenPrime(dupData)
+      , response = await simpleTokenPrimeObj.perform()
     ;
     assert.equal(response.isSuccess(), false);
   });
@@ -78,8 +78,8 @@ describe('services/transaction/transfer/simple_token', function () {
     var dupData = JSON.parse(JSON.stringify(testValidData));
     dupData.options.tag = 'a@b';
 
-    var simpleTokenObj = new platformServices.simpleToken(dupData)
-      , response = await simpleTokenObj.perform()
+    var simpleTokenPrimeObj = new platformServices.simpleTokenPrime(dupData)
+      , response = await simpleTokenPrimeObj.perform()
     ;
     assert.equal(response.isSuccess(), false);
   });
@@ -90,8 +90,8 @@ describe('services/transaction/transfer/simple_token', function () {
     var dupData = JSON.parse(JSON.stringify(testValidData));
     dupData.sender_name = 'Google';
 
-    var simpleTokenObj = new platformServices.simpleToken(dupData)
-      , response = await simpleTokenObj.perform()
+    var simpleTokenPrimeObj = new platformServices.simpleTokenPrime(dupData)
+      , response = await simpleTokenPrimeObj.perform()
     ;
     assert.equal(response.isSuccess(), false);
   });
@@ -101,8 +101,8 @@ describe('services/transaction/transfer/simple_token', function () {
     dupData.sender_name = ''; // has higher priority
     dupData.sender_address = 'abc';
 
-    var simpleTokenObj = new platformServices.simpleToken(dupData)
-      , response = await simpleTokenObj.perform()
+    var simpleTokenPrimeObj = new platformServices.simpleTokenPrime(dupData)
+      , response = await simpleTokenPrimeObj.perform()
     ;
     assert.equal(response.isSuccess(), false);
   });
@@ -110,10 +110,10 @@ describe('services/transaction/transfer/simple_token', function () {
   it('should fail when sender address is valid, but has no balance', async function () {
     var dupData = JSON.parse(JSON.stringify(testValidData));
     dupData.sender_name = ''; // has higher priority
-    dupData.sender_address = '0xb4d7bedf714e6c7cd1a641f705870fa19144a021';
+    dupData.sender_address = '0xb4d7bedf714e6c7cd1a641f705870fa19144a061';
 
-    var simpleTokenObj = new platformServices.simpleToken(dupData)
-      , response = await simpleTokenObj.perform()
+    var simpleTokenPrimeObj = new platformServices.simpleTokenPrime(dupData)
+      , response = await simpleTokenPrimeObj.perform()
     ;
     assert.equal(response.isSuccess(), false);
   });
@@ -123,8 +123,8 @@ describe('services/transaction/transfer/simple_token', function () {
     dupData.sender_name = ''; // has higher priority
     dupData.sender_passphrase = 'abc';
 
-    var simpleTokenObj = new platformServices.simpleToken(dupData)
-      , response = await simpleTokenObj.perform()
+    var simpleTokenPrimeObj = new platformServices.simpleTokenPrime(dupData)
+      , response = await simpleTokenPrimeObj.perform()
     ;
     assert.equal(response.isSuccess(), false);
   });
@@ -136,8 +136,8 @@ describe('services/transaction/transfer/simple_token', function () {
     dupData.recipient_name = ''; // has higher priority
     dupData.recipient_address = 'abc';
 
-    var simpleTokenObj = new platformServices.simpleToken(dupData)
-      , response = await simpleTokenObj.perform()
+    var simpleTokenPrimeObj = new platformServices.simpleTokenPrime(dupData)
+      , response = await simpleTokenPrimeObj.perform()
     ;
     assert.equal(response.isSuccess(), false);
   });
@@ -146,8 +146,8 @@ describe('services/transaction/transfer/simple_token', function () {
     var dupData = JSON.parse(JSON.stringify(testValidData));
     dupData.recipient_name = 'abc';
 
-    var simpleTokenObj = new platformServices.simpleToken(dupData)
-      , response = await simpleTokenObj.perform()
+    var simpleTokenPrimeObj = new platformServices.simpleTokenPrime(dupData)
+      , response = await simpleTokenPrimeObj.perform()
     ;
     assert.equal(response.isSuccess(), false);
   });
@@ -158,8 +158,8 @@ describe('services/transaction/transfer/simple_token', function () {
     var dupData = JSON.parse(JSON.stringify(testValidData));
     dupData.amount_in_wei = undefined;
 
-    var simpleTokenObj = new platformServices.simpleToken(dupData)
-      , response = await simpleTokenObj.perform()
+    var simpleTokenPrimeObj = new platformServices.simpleTokenPrime(dupData)
+      , response = await simpleTokenPrimeObj.perform()
     ;
     assert.equal(response.isSuccess(), false);
   });
@@ -168,8 +168,8 @@ describe('services/transaction/transfer/simple_token', function () {
     var dupData = JSON.parse(JSON.stringify(testValidData));
     dupData.amount_in_wei = 'abc';
 
-    var simpleTokenObj = new platformServices.simpleToken(dupData)
-      , response = await simpleTokenObj.perform()
+    var simpleTokenPrimeObj = new platformServices.simpleTokenPrime(dupData)
+      , response = await simpleTokenPrimeObj.perform()
     ;
     assert.equal(response.isSuccess(), false);
   });
@@ -178,8 +178,8 @@ describe('services/transaction/transfer/simple_token', function () {
     var dupData = JSON.parse(JSON.stringify(testValidData));
     dupData.amount_in_wei = 100.2;
 
-    var simpleTokenObj = new platformServices.simpleToken(dupData)
-      , response = await simpleTokenObj.perform()
+    var simpleTokenPrimeObj = new platformServices.simpleTokenPrime(dupData)
+      , response = await simpleTokenPrimeObj.perform()
     ;
     assert.equal(response.isSuccess(), false);
   });
@@ -188,8 +188,8 @@ describe('services/transaction/transfer/simple_token', function () {
     var dupData = JSON.parse(JSON.stringify(testValidData));
     dupData.amount_in_wei = 0.2;
 
-    var simpleTokenObj = new platformServices.simpleToken(dupData)
-      , response = await simpleTokenObj.perform()
+    var simpleTokenPrimeObj = new platformServices.simpleTokenPrime(dupData)
+      , response = await simpleTokenPrimeObj.perform()
     ;
     assert.equal(response.isSuccess(), false);
   });
@@ -198,8 +198,8 @@ describe('services/transaction/transfer/simple_token', function () {
     var dupData = JSON.parse(JSON.stringify(testValidData));
     dupData.amount_in_wei = 0;
 
-    var simpleTokenObj = new platformServices.simpleToken(dupData)
-      , response = await simpleTokenObj.perform()
+    var simpleTokenPrimeObj = new platformServices.simpleTokenPrime(dupData)
+      , response = await simpleTokenPrimeObj.perform()
     ;
     assert.equal(response.isSuccess(), false);
   });
@@ -208,8 +208,8 @@ describe('services/transaction/transfer/simple_token', function () {
     var dupData = JSON.parse(JSON.stringify(testValidData));
     dupData.amount_in_wei = -100;
 
-    var simpleTokenObj = new platformServices.simpleToken(dupData)
-      , response = await simpleTokenObj.perform()
+    var simpleTokenPrimeObj = new platformServices.simpleTokenPrime(dupData)
+      , response = await simpleTokenPrimeObj.perform()
     ;
     assert.equal(response.isSuccess(), false);
   });
@@ -219,8 +219,8 @@ describe('services/transaction/transfer/simple_token', function () {
   it('should pass when everything is valid', async function () {
     var dupData = JSON.parse(JSON.stringify(testValidData));
 
-    var simpleTokenObj = new platformServices.simpleToken(dupData)
-      , response = await simpleTokenObj.perform()
+    var simpleTokenPrimeObj = new platformServices.simpleTokenPrime(dupData)
+      , response = await simpleTokenPrimeObj.perform()
     ;
     assert.equal(response.isSuccess(), true);
     assert.isNotNull(response.data.transaction_uuid);
@@ -231,8 +231,8 @@ describe('services/transaction/transfer/simple_token', function () {
     var dupData = JSON.parse(JSON.stringify(testValidData));
     dupData.options.returnType = 'myReturnType';
 
-    var simpleTokenObj = new platformServices.simpleToken(dupData)
-      , response = await simpleTokenObj.perform()
+    var simpleTokenPrimeObj = new platformServices.simpleTokenPrime(dupData)
+      , response = await simpleTokenPrimeObj.perform()
     ;
 
     assert.equal(response.isSuccess(), true);
@@ -245,8 +245,8 @@ describe('services/transaction/transfer/simple_token', function () {
     var dupData = JSON.parse(JSON.stringify(testValidData));
     dupData.options.returnType = 'uuid';
 
-    var simpleTokenObj = new platformServices.simpleToken(dupData)
-      , response = await simpleTokenObj.perform()
+    var simpleTokenPrimeObj = new platformServices.simpleTokenPrime(dupData)
+      , response = await simpleTokenPrimeObj.perform()
     ;
 
     assert.equal(response.isSuccess(), true);
@@ -259,8 +259,8 @@ describe('services/transaction/transfer/simple_token', function () {
     var dupData = JSON.parse(JSON.stringify(testValidData));
     dupData.options.returnType = 'txReceipt';
 
-    var simpleTokenObj = new platformServices.simpleToken(dupData)
-      , response = await simpleTokenObj.perform()
+    var simpleTokenPrimeObj = new platformServices.simpleTokenPrime(dupData)
+      , response = await simpleTokenPrimeObj.perform()
     ;
 
     assert.equal(response.isSuccess(), true);
@@ -268,5 +268,8 @@ describe('services/transaction/transfer/simple_token', function () {
     assert.isNotNull(response.data.transaction_hash);
     assert.isNumber(response.data.transaction_receipt.blockNumber);
   });
+
+  //TODO: check balance in cache
+
 
 });
