+++ test/integration/services/utils_generate_address.js
@@ -3,7 +3,7 @@ const chai = require('chai')
   , assert = chai.assert;
 
 // Load cache service
-const rootPrefix = "../.."
+const rootPrefix = "../../.."
   , openstPlatform = require(rootPrefix + '/index')
   , platformServices = openstPlatform.services.utils
   , web3ProviderFactory = require(rootPrefix + '/lib/web3/providers/factory')
